% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{desc}
\alias{desc}
\alias{desc_dev_deps}
\alias{desc_maintainer}
\title{Return the number of development dependencies in a package description file.}
\usage{
desc_dev_deps(description)

desc_maintainer(description)
}
\arguments{
\item{description}{A [desc::desc] object.}
}
\description{
Return the number of development dependencies in a package description file.
}
