% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_org_name}
\alias{get_org_name}
\title{Get org name}
\usage{
get_org_name(org)
}
\arguments{
\item{org}{Org or user login}
}
\value{
string
}
\description{
Get org name
}
\examples{
get_org_name("r-lib")
get_org_name("jimhester")
}
