% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_progress.R
\name{issue_progress}
\alias{issue_progress}
\title{Retrieve opened and closed issues and pull requests since a given date}
\usage{
issue_progress(org, start = today() - dweeks(1))
}
\arguments{
\item{org}{org to query}

\item{start}{Datetime to start query from}
}
\description{
Retrieve opened and closed issues and pull requests since a given date
}
