% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{normalize_privacy}
\alias{normalize_privacy}
\title{Normalize the repository privacy}
\usage{
normalize_privacy(privacy = c("PUBLIC", "PRIVATE", "BOTH"))
}
\arguments{
\item{privacy}{Github repository privacy}
}
\description{
Normalize the repository privacy
}
