% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{num_downloads}
\alias{num_downloads}
\title{Number of downloads from RStudio mirror}
\usage{
num_downloads(package, when = "last-week")
}
\arguments{
\item{package}{One or more packages to search}

\item{when}{\code{last-day}, \code{last-week} or \code{last-month} (see 
details). If this is given, then \code{from} and \code{to} are ignored.}
}
\description{
Number of downloads from RStudio mirror
}
