% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{org_data}
\alias{org_data}
\title{Compute an organization summary}
\usage{
org_data(org)
}
\arguments{
\item{org}{A GitHub user, either a normal user or an organization}
}
\description{
Compute an organization summary
}
