% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequests.R
\name{org_pr}
\alias{org_pr}
\title{Compute an pull request reviewer summary}
\usage{
org_pr(org)
}
\arguments{
\item{org}{A GitHub user, either a normal user or an organization}
}
\description{
Compute an pull request reviewer summary
}
