% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{repo_status_badge}
\alias{repo_status_badge}
\title{Generate a repo status badge}
\usage{
repo_status_badge(owner, package)
}
\arguments{
\item{owner}{The repository owner}

\item{package}{The package}
}
\description{
Generate a repo status badge
}
