% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{reverse_dependencies}
\alias{reverse_dependencies}
\title{Calculate number or reverse dependencies}
\usage{
reverse_dependencies(package)
}
\arguments{
\item{package}{One or more packages to search}
}
\description{
Calculate number or reverse dependencies
}
