% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparkline_table}
\alias{sparkline_table}
\title{Plot a sparkline table}
\usage{
sparkline_table(data, sparkline_column, ...)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{sparkline_column}{The column to convert to a sparkline}
}
\description{
Plot a sparkline table
}
