% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{status_badge}
\alias{status_badge}
\alias{travis_status_badge}
\alias{appveyor_status_badge}
\title{Generate a status badge}
\usage{
travis_status_badge(owner, package)

appveyor_status_badge(owner, package)
}
\arguments{
\item{owner}{The repository owner}

\item{package}{The package}
}
\description{
Generate a status badge
}
