% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversedashboard-package.R
\docType{package}
\name{tidyversedashboard-package}
\alias{tidyversedashboard}
\alias{tidyversedashboard-package}
\title{tidyversedashboard: Tidyverse activity dashboard}
\description{
A dashboard to show repository activity in tidyverse organisations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimhester/tidyversedashboard}
  \item Report bugs at \url{https://github.com/jimhester/tidyversedashboard/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com}

}
\keyword{internal}
