% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{weekly_issues}
\alias{weekly_issues}
\title{Retrieve opened and closed issues and pull requests since a given date}
\usage{
weekly_issues(org, start = today() - dweeks(1))
}
\arguments{
\item{org}{org to query}

\item{start}{Datetime to start query from}
}
\description{
Retrieve opened and closed issues and pull requests since a given date
}
