% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_org_logo}
\alias{get_org_logo}
\title{Get org logo}
\usage{
get_org_logo(org)
}
\arguments{
\item{org}{}
}
\value{
writes the avatar as a local file logo.png
}
\description{
Get org logo
}
\examples{
get_org_logo("r-lib")
}
