% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{status_badge}
\alias{status_badge}
\alias{github_status_badge}
\alias{cran_status_badge}
\alias{coverage_status_badge}
\title{Generate a status badge}
\usage{
github_status_badge(owner, package)

cran_status_badge(package)

coverage_status_badge(owner, package, default = "main")
}
\arguments{
\item{owner}{The repository owner}

\item{package}{The package}

\item{default}{Main branch of the git repository}
}
\description{
Generate a status badge
}
