% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{weekly_commits}
\alias{weekly_commits}
\title{Lookup weekly commit counts for each repo}
\usage{
weekly_commits(org, package)
}
\arguments{
\item{org}{A GitHub user, either a normal user or an organization}

\item{package}{One or more packages to search}
}
\description{
Lookup weekly commit counts for each repo
}
